add the feature gates LoggingAlphaOptions and LoggingBetaOptions
These two feature gates come from Kubernetes. It is necessary to add
these two feature gates to enable the use of the flags

  --log-json-split-stream
  --log-json-info-buffer-size quantity

Otherwise, these flags would show in the --help but not be usable. The
alternative would have been to "remove" these flags, but we chose to
follow Kubernetes' practice in terms of flags.

Signed-off-by: Maël Valais <mael@vls.dev>